+++ src/main/java/app/BuddyRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(path = "buddyInfo")
 public interface BuddyRepository extends CrudRepository<BuddyInfo, Long> {
-    BuddyInfo findById(long id);
+    BuddyInfo findById(int id);
 }
